requirements_exe.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
## !! Could not determine repository location
-e /Users/home/Documents/PyProjects/FinanceGPT2
fonttools==4.55.0
fpdf2==2.8.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
json5==0.9.11
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.14.2
numpy==1.26.4
openai==1.55.2
packaging==23.2
pandas==2.1.1
pdfkit==1.0.0
Pillow==9.5.0
plotly==5.15.0
pluggy==1.5.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
Pympler==1.1
pyparsing==3.2.0
PyPDF2==3.0.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.24.0
tenacity==8.2.2
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.1.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
yarl==1.18.0
zipp==3.21.0
